+++ modules/canvaskit/tests/path.spec.js
@@ -216,4 +216,49 @@ describe('CanvasKit\'s Path Behavior', function() {
             reportSurface(surface, 'arcto_path', done);
         }));
     });
+
+    it('can draw a path using relative functions', function(done) {
+        LoadCanvasKit.then(catchException(done, () => {
+            // This is taken from example.html
+            const surface = CanvasKit.MakeCanvasSurface('test');
+            expect(surface).toBeTruthy('Could not make surface')
+            if (!surface) {
+                done();
+                return;
+            }
+            const canvas = surface.getCanvas();
+            const paint = new CanvasKit.SkPaint();
+            paint.setStrokeWidth(1.0);
+            paint.setAntiAlias(true);
+            paint.setColor(CanvasKit.Color(0, 0, 0, 1.0));
+            paint.setStyle(CanvasKit.PaintStyle.Stroke);
+
+            const path = new CanvasKit.SkPath();
+            path.rMoveTo(20, 5)
+                .rLineTo(10, 15)  // 30, 20
+                .rLineTo(10, -5);  // 40, 10
+            path.rLineTo(10, 10);  // 50, 20
+            path.rLineTo(10, -20); // 60, 0
+            path.rLineTo(-40, 5);  // 20, 5
+
+            path.moveTo(20, 80)
+                .rCubicTo(70, -70, 140, 70, 170, -70); // 90, 10, 160, 150, 190, 10
+
+            path.moveTo(36, 148)
+                .rQuadTo(30, 40, 84, -12) // 66, 188, 120, 136
+                .lineTo(36, 148);
+
+            path.moveTo(150, 180)
+                .rArcTo(24, 24, 45, true, false, -68, -24); // 82, 156
+            path.lineTo(160, 160);
+
+            canvas.drawPath(path, paint);
+
+            surface.flush();
+            path.delete();
+            paint.delete();
+
+            reportSurface(surface, 'path_relative', done);
+        }));
+    });
 });